shark is now red
previously, the shark was black.
now, the shark is red
next, the ability to toggle the presence of the shark
and fix the flocking boids algorithm as it doesn't respect the seperation, alignment and cohesion rules as expected. boids are jittery and some boids are removed after restarting the simulation